+++ app/public/addPet.html
@@ -105,10 +105,10 @@
           </select>
         </div>
         <!-- img choose -->
-        <form>
+        <form enctype="multipart/form-data">
           <div class="form-group">
             <label for="pictureName">Animal's picture</label>
-            <input type="file" class="form-control-file getPictName" id="pictureName" accept="image/*">
+            <input name="petPicture" type="file" class="form-control-file getPictName" id="pictureName" accept="image/*">
           </div>
         </form>
         <!-- <div class="input-group mb-3">
